Remove behavior block after associated trait is removed
If a behavior block has a trait associated with it (and is applied),
i.e. the behavior input has been removed, AND is waiting for a trait
block to be dropped or a trait block has already been dropped in the
behavior block, and then the trait is removed from the breed via the
species inspector panel, the behavior block is now meaningless and
must be removed.

For the time being this will work if only one trait is associated
with a behavior block because the code checks applicableTraits,
not APPLIED traits. Perhaps another list should be added to
behavior block to keep a track of which traits have been applied to
the block.