Fixed GoogleDoc Bug#13.
If there are behavior blocks in a breed block, and a new trait is defined,
only the behavior blocks to which the trait applies are updated. Other
behavior blocks are unaffected.

Similarly, if a breed has a trait, a behavior block (to which the trait
applies) is added, but the trait block is not yet dropped (the behavior
block is waiting for the trait), and then the trait is removed from the
breed block, this behavior block is removed (eventhough the trait hasn't
yet been applied). As this the removed trait no longer exists, there is
reason for the behavior block to wait for the trait. Removing the block
may be easier than making the behvior input visible and removing the
'add trait block here panel'.