Fixed GoogleDoc bug#32
If a trait is defined, move block added (or other way round), i.e.
move (behavior) block is waiting for trait (block), and if you open
species inspector and click okay, the "add yellow block here" panel
doesn't get repeatedly added. If the block is waiting for a trait,
then there is no need to add the panel or remove the behavior input.