Fixed buglist #45 -- pen not made error
The getYLabel funtion in quantitiy block was referencing tBlock even when
a traitblock is not applicable. This was causing a null pointer exception
and the plot pen for the NetLogo plot was not getting created.

Now there is a check for tBlock. Also, if a PrettyInput called "breed-type"
exists then the corresponding breed name is attached to the yLabel.