Enabled mutate implementation.
An empty mutate procedure is added if a reproduce block is used. The NetLogo
code for the reproduce block always calls the mutate procedure. If mutation
is not enabled in the library XML then no mutation related code is generate.
If mutation is enabled, then the corresponding mutation related code is
generated for all breeds and all traits.