Fixed condition drop bug
Fixed bug that prevented trait block being dropped into a condition block
if a behavior block was already present. This was due to the 1 block max
limit in condition block. The limit was getting enforced on the trait block
as well. Fixed by processing traitblock before checking for the limit.