Fixed wrong trait block in quantity block bug
If two species have the same trait, there are two trait blocks, but
when a model is opened only one of the trait blocks is gets added to
quantity block (same should happen with behavior or condition block).
This would happen because the function getTrait() only required
the trait name. Now the function requires trait and breed name to pick
the correct trait block to return.