Misc procedures related to make-distribution no longer throw errors
The procedure code for misc procedure is populated only if a trait is
defined. If no trait is defined, the procedure code (body of the procedure)
is blank, hence no error about undefined species or traits is thrown.

Also, the make distribution button doesn't appear until at least one
trait is defined.