Update DeltaTickFormat.dtd
Change breed, a content of library object, to be required rather than optional.
"*" means it's ok if no breed is defined, but this would actually break DT and the library won't be loaded. Instead I used "+" instead because this makes clear that at least one MUST be present, and more than one could be present if needed.